+++ docs/conf.py
@@ -40,15 +40,15 @@ sys.path.append(os.path.abspath("../sphinxcontrib"))
 # built documents.
 #
 # The short X.Y version.
-version = "0.7"
+version = "1.0"
 # The full version, including alpha/beta/rc tags.
-release = "0.7.9"
+release = "1.0.1"
 
 on_rtd = os.environ.get("READTHEDOCS") == "True"
 
 extensions = [
-    "sphinxcontrib.needs",
     "sphinxcontrib.plantuml",
+    "sphinx_needs",
     "sphinx.ext.autodoc",
     "matplotlib.sphinxext.plot_directive",
     "sphinx_copybutton",
@@ -150,6 +150,7 @@ DEFAULT_DIAGRAM_TEMPLATE = (
 
 # needs_template = TITLE_TEMPLATE
 # needs_diagram_template = DEFAULT_DIAGRAM_TEMPLATE
+
 
 needs_types = [
     dict(directive="req", title="Requirement", prefix="R_", color="#BFD8D2", style="node"),
@@ -342,6 +343,15 @@ needs_string_links = {
     },
 }
 
+# needs_external_needs = [
+#     {
+#         "base_url": "https://sphinxcontrib-needs.readthedocs.io/en/latest",
+#         "json_path": "examples/needs.json",
+#         "id_prefix": "EXT_",
+#         "css_class": "external_link",
+#     },
+# ]
+
 # build needs.json to make permalinks work
 needs_build_json = True
 
@@ -492,6 +502,7 @@ html_theme_options = {
     ],
     "toc_title_is_page_title": True
 }
+
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,