+++ sphinx_needs/services/base.py
@@ -1,0 +1,12 @@
+from sphinx_needs.logging import get_logger
+
+
+class BaseService:
+    def __init__(self, *args, **kwargs):
+        self.log = get_logger(__name__)
+
+    def request(self, *args, **kwargs):
+        raise NotImplementedError("Must be implemented by the service!")
+
+    def debug(self, *args, **kwargs):
+        raise NotImplementedError("Must be implemented by the service!")